tm2:latest

+++ start.sh
@@ -14,12 +14,12 @@ then
    unzip /srv/ManiaPlanetBetaServer_2015-06-16.zip -d /srv/
 fi
 
-# if [ ! -f /srv/UserData/Config/dedicated_cfg.default.txt ]
-# then
-#     cp /srv/UserData/Config/dedicated_cfg.default.txt /srv/UserData/Config/dedicated_cfg 
-# fi
+if [ -f /srv/UserData/Config/dedicated_cfg.default.txt ]
+then
+    cp /srv/UserData/Config/dedicated_cfg.default.txt /srv/UserData/Config/dedicated_cfg 
+fi
 
-cp /srv/UserData/Config/dedicated_cfg.default.txt /srv/UserData/Config/dedicated_cfg
+# cp /srv/UserData/Config/dedicated_cfg.default.txt /srv/UserData/Config/dedicated_cfg
 
 if [ ! -f /srv/RunSrvTM.sh ]
 then
